requirements.txt
opencv-python
numpy
flask
tensorflow
pillow